test(projects): the merge refusal takes its own file
project_integration_test.go crossed the 1000-line test cap once this
branch's case landed on top of what main had grown.

The merge refusal is a rule of its own — two companies each carrying
live work do not combine — and it is read from two ends: the engine that
refuses the write, and the card that decides whether to offer the merge.
Those six tests move together, and liveProjectEdge's "Held by" pointer
follows them, which is what the uniqueness-claim gate caught.
